moved items around to highlight importance
I tried putting little GitHub and LinkedIn and Google Scholar logos in
place of the bullets in the Social Networks section.  It worked, using
the syntax \item[{\includegraphics[height=\baselineskip]{githib.png}}
but the result looked cluttered whether I put the logo at the beginning
or the end of the line.