remove degree dates and street address
On the advice of someone to avoid ageism, I removed the years when
degrees were conferred.  I removed the street address as irrelevant,
leaving only city and state, them moved email addresses higher than
telephone number because they are now relatively more important.  With
the extra space thereby gained, I moved proposal wins and funding to the
first page, then moved the security clearance to the top of the second
page.  The last page is now a bit sparse but there is room in case I
want to add anything else.

I am worried that the work experience section is weak and doesn't have
enough meat in it.