added a QR code to the header
I am not sure the QR code is a great idea or not.  It currently contains
only a minimal name and email address vCard.  I would like it to contain
a URL to the GitHub page, but that increases the size of the QR code,
making it less likely to scan.  I may re-add it later, now that I have
the formatting working in LaTeX.