changed to 12pt to avoid ligatures problem
There is a problem with ligatures in 11pt size.  Ligatures and the CMAP
work fine at 10pt and 12pt but not at 11pt.  I can't find the problem in
the 11pt compilation so to get it working for now, I reformatted the
text to 12pt which is probably easier for hiring managers to read
anyway.  10pt is too small.