added a sheet of references; fixed mailto: links
Makefile now builds two PDF files, CV and references. Also, when using
\href{}, don't use \url{} in the right hand side; use \nolinkurl{}
instead. It looks the same, but the first way tickles a bug that causes
mailto: links to become relative http: links in the generated PDF file.
Finally, I added a spell target to the Makefile for references sheet. Ed
Telders' references are not up to date yet.